pyproject.toml
[tool.poetry]
name = "download-social-bot"
version = "0.1.0"
description = "Donwload bot"
authors = ["eastsidekillas"]
license = "Private"
packages = [{include = "*"}]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12.0, <3.13.0"
yt-dlp = "^2025.9.26"
youtube-dl = "^2021.12.17"
requests = "^2.32.5"
bs4 = "^0.0.2"
shazamio = "^0.8.1"
aiogram = "^3.22.0"
python-decouple = "^3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
